Introduce lockscreen clock/date styles [1/2]
Inspired by OnePlus and their recent options to switch lockscreen
clock styles on the OP6, I took it upon myself to do the same with our ROM.

This isn't as fancy as OnePlus' because is just using a list preference instead
of adding in all the eye candy but it gets the job done!

Below is a run down of what took place...

- Mirrored the analog clock class to avoid issues with any apps that may
or may not be utilizing. Google deprecated this class back in 2015 but there's
still a chance that something may go wrong some where. Better safe than sorry.

https://github.com/DirtyUnicorns/android_frameworks_base/blob/o8x/core/java/android/widget/AnalogClock.java

- Added the analog clock to the keyguard and made it optional
- Manipulated the clockview format for both 12 and 24 hour to achieved the 'sammy' style
- Made the sammy style optional and created a switch case for all 3.
- Set padding, singeline and other layout changes programmatically to minimize the jank
- Increased the date widget size to 20sp becuase it just looks better
- Reworked/cleaned up some existing code (Huge thanks to @ezio84)
- Created bold sammy by using html formatting, based off the sammy one
- Created bold digital by using html formatting, based off the digital one
- Created some date styles to match clock styles
- Vectorized the analog clock images so we can theme
- Modified Analog clock class to fix ambient display issues (@ezio84)
- Added DeadPool analog clock because Fuck you, that's why! lol

For right now there's six clock styles and 3 date styles:

Clock styles---
- Default
- Digital (bold)
- Analog
- Analog (deadpool)
- Sammy
- Sammy (bold)

Date styles----
- Default
- Semi transparent box
- Semi transparent box (round corners)

Change-Id: I3122e832a7f320cedaa8ec7b336d0e0c767f94e4
Signed-off-by: Shubham Singh <coolsks94@gmail.com>